refactor: split channel presence policy
